+++ internal/api/activitypub/users/user_test.go
@@ -33,7 +33,6 @@ import (
 	"github.com/superseriousbusiness/gotosocial/internal/state"
 	"github.com/superseriousbusiness/gotosocial/internal/storage"
 	"github.com/superseriousbusiness/gotosocial/internal/typeutils"
-	"github.com/superseriousbusiness/gotosocial/internal/webpush"
 	"github.com/superseriousbusiness/gotosocial/testrig"
 )
 
@@ -105,7 +104,7 @@ func (suite *UserStandardTestSuite) SetupTest() {
 		&suite.state,
 		suite.federator,
 		suite.emailSender,
-		webpush.NewNoopSender(),
+		testrig.NewNoopWebPushSender(),
 		suite.mediaManager,
 	)
 	testrig.StartWorkers(&suite.state, suite.processor.Workers())
