+++ internal/api/client/exports/exports_test.go
@@ -34,7 +34,6 @@ import (
 	"github.com/superseriousbusiness/gotosocial/internal/oauth"
 	"github.com/superseriousbusiness/gotosocial/internal/state"
 	"github.com/superseriousbusiness/gotosocial/internal/typeutils"
-	"github.com/superseriousbusiness/gotosocial/internal/webpush"
 	"github.com/superseriousbusiness/gotosocial/testrig"
 )
 
@@ -96,7 +95,7 @@ func (suite *ExportsTestSuite) SetupTest() {
 		&suite.state,
 		federator,
 		testrig.NewEmailSender("../../../../web/template/", nil),
-		webpush.NewNoopSender(),
+		testrig.NewNoopWebPushSender(),
 		mediaManager,
 	)
