+++ internal/api/client/import/import_test.go
@@ -32,7 +32,6 @@ import (
 	"github.com/superseriousbusiness/gotosocial/internal/oauth"
 	"github.com/superseriousbusiness/gotosocial/internal/state"
 	"github.com/superseriousbusiness/gotosocial/internal/typeutils"
-	"github.com/superseriousbusiness/gotosocial/internal/webpush"
 	"github.com/superseriousbusiness/gotosocial/testrig"
 )
 
@@ -93,7 +92,7 @@ func (suite *ImportTestSuite) SetupTest() {
 		&suite.state,
 		federator,
 		testrig.NewEmailSender("../../../../web/template/", nil),
-		webpush.NewNoopSender(),
+		testrig.NewNoopWebPushSender(),
 		mediaManager,
 	)
 	testrig.StartWorkers(&suite.state, processor.Workers())
