+++ internal/api/client/reports/reports_test.go
@@ -31,7 +31,6 @@ import (
 	"github.com/superseriousbusiness/gotosocial/internal/state"
 	"github.com/superseriousbusiness/gotosocial/internal/storage"
 	"github.com/superseriousbusiness/gotosocial/internal/typeutils"
-	"github.com/superseriousbusiness/gotosocial/internal/webpush"
 	"github.com/superseriousbusiness/gotosocial/testrig"
 )
 
@@ -96,7 +95,7 @@ func (suite *ReportsStandardTestSuite) SetupTest() {
 		&suite.state,
 		suite.federator,
 		suite.emailSender,
-		webpush.NewNoopSender(),
+		testrig.NewNoopWebPushSender(),
 		suite.mediaManager,
 	)
 	suite.reportsModule = reports.New(suite.processor)
